further splitting tidy
